Fixing rebase so changes are only to elasticsearch directory

+++ fluentd/run.sh
@@ -17,13 +17,12 @@ done
 fluentdargs="--no-supervisor"
 if [[ $VERBOSE ]]; then
   set -ex
-  fluentdargs="-vv"
+  fluentdargs="$fluentdargs -vv"
   echo ">>>>>> ENVIRONMENT VARS <<<<<"
   env | sort
   echo ">>>>>>>>>>>>><<<<<<<<<<<<<<<<"
 else
   set -e
-  fluentdargs=
 fi
 
 #NOTE: USE_CRIO variable used in generate_throttle_configs.rb as well
@@ -285,7 +284,8 @@ if [[ "${USE_REMOTE_SYSLOG:-}" = "true" ]] ; then
     # The symlink is a workaround for https://github.com/openshift/origin-aggregated-logging/issues/604
     found=
     for file in /usr/share/gems/gems/fluent-plugin-remote-syslog-*/lib/fluentd/plugin/*.rb ; do
-        if [ -f "$file" ] ; then
+        bname=$(basename $file)
+        if [ ! -e "/etc/fluent/plugin/$bname" -a -f "$file" ] ; then
             ln -s $file /etc/fluent/plugin/
             found=true
         fi
@@ -293,7 +293,8 @@ if [[ "${USE_REMOTE_SYSLOG:-}" = "true" ]] ; then
     if [ -z "${found:-}" ] ; then
         # not found in rpm location - look in alternate location
         for file in /opt/app-root/src/gems/fluent-plugin-remote-syslog*/lib/fluentd/plugin/*.rb ; do
-            if [ -f "$file" ] ; then
+            bname=$(basename $file)
+            if [ ! -e "/etc/fluent/plugin/$bname" -a -f "$file" ] ; then
                 ln -s $file /etc/fluent/plugin/
             fi
         done
